fix: re-set-up multimedia (#228)
* fix: specify only custom config file

When we specify both an upstream defconfig and a local custom
full-config, buildroot gets very confused and seems to use neither of
them, which I think we can all agree isn't right. This is why none of
our modules seemed to be there.

* fix: audio modules, alsa config for sound output

Before the update, I think we had a pretty simple audio setup and after
the update we were missing the driver modules for audio on our soc.
Bringing the driver modules back is pretty easy, but we now also need a
bunch of extra alsa configuration to make sure that the volume is at
100% because at 50% it's completely inaudible.

We can use amixer (NOT alsamixer, that's a TUI interactive mixer) to set
the volume high, and alsactl to store it, but we also need to make sure
we have somewhere for alsactl's state file, and then bam we've got
persistent volume across reboots.

* fix: reenable video capture

Another set of kernel modules that got lost were the ones for the
drivers that let us capture video from the camera. This re-adds them.